refactor(runtime): split skill observation boundaries
Keep the stable skill_load_observation import surface while separating closed
protocol validation from executable snapshot and process lifecycle ownership.
Each production module now has one cohesive responsibility and remains below
250 nonblank, non-comment lines.

Constraint: preserve the public import/API and all fail-closed protocol, snapshot, and lifecycle behavior.
Rejected: leave the 600-line mixed-responsibility module intact | it exceeds the release size rule and obscures security boundaries.
Confidence: high
Scope-risk: narrow
Directive: keep protocol parsing, process execution, and public orchestration in their current cohesive modules.
Tested: 56 focused observation, cleanup, Windows, replay, CLI, and safety-contract tests; Ruff; compileall; basedpyright 0 errors with warnings reduced from the prior module result.
Not-tested: native Windows execution and a real observed Hermes inventory.
Omo-Task: st_01a0403a

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
