+++ tests/test_handoff_safety_contract_enforcement.py
@@ -146,9 +146,10 @@ PROCESS_SPAWN_ALLOWLIST: dict[str, str] = {
         "operator-only, explicitly confirmed `ask_before_dispatch` seam for one bounded local "
         "`hermes --oneshot --model` child; suppresses recursion and cleans its process group."
     ),
-    "src/coding/skill_load_observation.py": (
-        "operator-only `omh coding hermes-child skill-load-probe --confirm-dispatch`; runs one "
-        "bounded local machine-inventory command, never a model or skill body."
+    "src/coding/skill_load_process.py": (
+        "private executable snapshot and process lifecycle adapter for the operator-only "
+        "`omh coding hermes-child skill-load-probe --confirm-dispatch`; runs one bounded local "
+        "machine-inventory command, never a model or skill body."
     ),
     "src/coding/_hermes_child_process.py": (
         "private lifecycle helper for the explicitly confirmed Hermes child seam; relays signals, "
